+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - ScrollQueue - Added OnItemTrimmed event
  - ScrollQueueTests - Added OnItemTrimmed event test
+ - DateTimeNullableEventArgs
 
 ## [11.0.0] - 2020-03-20
 ### Added