+++ CHANGELOG.md
@@ -5,10 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [11.1.0] - 2020-05-19
 ### Added
  - ScrollQueue - Added OnItemTrimmed event
- - ScrollQueueTests - Added OnItemTrimmed event test
- - DateTimeNullableEventArgs
+ - Added DateTimeNullableEventArgs
 
 ## [11.0.0] - 2020-03-20
 ### Added